requirements.txt
certifi==2018.10.15
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.3
django-rest-auth==0.9.3
django-rest-swagger==2.2.0
djangorestframework==3.9.0
djangorestframework-simplejwt==3.3
idna==2.7
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
netifaces==0.10.7
openapi-codec==1.3.2
psycopg2==2.7.6.1
PyJWT==1.6.4
pytz==2018.7
PyYAML==3.13
requests==2.20.1
simplejson==3.16.0
six==1.11.0
uritemplate==3.0.0
urllib3==1.24.1
